# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Router Console and OpenAI-compatible model gateway API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models:
    get:
      operationId: listModels
      tags: [console]
      summary: List available models
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/Model"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    post:
      operationId: createModel
      tags: [console]
      summary: Add a provider model
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ModelInput"}
      responses:
        "201":
          description: Model created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Model"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
...
  /v1/models:
    get:
      operationId: gatewayListModels
      tags: [gateway]
      summary: OpenAI-compatible model list
      security: [{ApiKey: []}]
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OpenAIModelList"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /v1/chat/completions:
    post:
      operationId: chatCompletion
      tags: [gateway]
      summary: Create an OpenAI-compatible chat completion
      security: [{ApiKey: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ChatCompletionInput"}
      responses:
        "200":
          description: Chat completion (JSON or text/event-stream when stream is true)
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ChatCompletionOutput"}
            text/event-stream:
              schema: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "502": {$ref: "#/components/responses/ProviderError"}
        "503": {$ref: "#/components/responses/Unavailable"}
components:
  securitySchemes:
    ApiKey:
      type: http
      scheme: bearer
      description: Application API key created in the console
...
  responses:
    Unauthorized:
      description: Authentication required or API key invalid
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    ProviderError:
      description: Provider request failed
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unavailable:
      description: No provider is currently available
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      properties:
        error: {type: string}
        requestId: {type: ["string", "null"]}
      required: [error]
...
    Model:
      type: object
      properties:
        id: {type: string, format: uuid}
        providerId: {type: string, format: uuid}
        providerName: {type: string}
        name: {type: string}
        modelName: {type: string}
        contextWindow: {type: integer}
        inputPricePerMillion: {type: number}
        outputPricePerMillion: {type: number}
        supportsStreaming: {type: boolean}
        supportsVision: {type: boolean}
        supportsTools: {type: boolean}
        enabled: {type: boolean}
      required: [id, providerId, providerName, name, modelName, contextWindow, inputPricePerMillion, outputPricePerMillion, supportsStreaming, supportsVision, supportsTools, enabled]
    ModelInput:
      type: object
      properties:
        providerId: {type: string, format: uuid}
        name: {type: string, minLength: 1, maxLength: 160}
        modelName: {type: string, minLength: 1, maxLength: 160}
        contextWindow: {type: integer, minimum: 1}
        inputPricePerMillion: {type: number, minimum: 0}
        outputPricePerMillion: {type: number, minimum: 0}
        supportsStreaming: {type: boolean}
        supportsVision: {type: boolean}
        supportsTools: {type: boolean}
      required: [providerId, name, modelName, contextWindow, inputPricePerMillion, outputPricePerMillion, supportsStreaming, supportsVision, supportsTools]
...
    ChatMessage:
      type: object
      properties:
        role: {type: string, enum: [system, user, assistant, tool]}
        content:
          oneOf:
            - {type: string}
            - {type: array, items: {type: object, additionalProperties: true}}
        name: {type: string}
        tool_call_id: {type: string}
      required: [role, content]
    ChatCompletionInput:
      type: object
      properties:
        model: {type: string, minLength: 1}
        messages: {type: array, minItems: 1, items: {$ref: "#/components/schemas/ChatMessage"}}
        temperature: {type: number, minimum: 0, maximum: 2}
        max_tokens: {type: integer, minimum: 1}
        top_p: {type: number, minimum: 0, maximum: 1}
        stream: {type: boolean, default: false}
        tools: {type: array, items: {type: object, additionalProperties: true}}
        tool_choice: {type: [string, object]}
        n: {type: integer, minimum: 1, maximum: 1}
      required: [model, messages]
    ChatCompletionOutput:
      type: object
      properties:
        id: {type: string}
        object: {type: string}
        created: {type: integer}
        model: {type: string}
        choices: {type: array, items: {type: object, additionalProperties: true}}
        usage:
          type: object
          properties:
            prompt_tokens: {type: integer}
            completion_tokens: {type: integer}
            total_tokens: {type: integer}
          required: [prompt_tokens, completion_tokens, total_tokens]
      required: [id, object, created, model, choices]
    OpenAIModelList:
      type: object
      properties:
        object: {type: string, const: list}
        data: {type: array, items: {type: object, additionalProperties: true}}
      required: [object, data]