# lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AI Router Console and OpenAI-compatible model gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface Error {
  error: string;
  /** @nullable */
  requestId?: string | null;
}